refactor: horas de clase son siempre enteros — eliminar decimales de DB, backend y frontend
- DB (local): Step 1a ejecutado — 69 asignaciones y 69 evaluacion_virtual_resultado redondeadas
- migrations/009_horas_integer.sql + run_migration_009.py: ALTER TABLE para servidor (requiere postgres)
- routers: catalogos, quincenas, evaluacion, portal, estadisticas — float() -> int() en horas
- Pydantic models: horas_semana, horas_bloque, horas_afectadas, horas_suplidas -> int
- scripts: importar_horarios_*.py — round(mins/60, 2) -> max(1, round(mins/60))
- frontend: Horarios.jsx y QuincenaDetalle.jsx — step=1, parseInt, quitar toFixed(1) en horas

+++ routers/evaluacion.py
@@ -174,7 +174,7 @@ async def listar_virtual(
 
     for asig in asignaciones:
         asig_id = asig['asignacion_id']
-        horas_quincena = float(asig['horas_semana'] or 0) * n_semanas
+        horas_quincena = int(asig['horas_semana'] or 0) * n_semanas
 
         cur.execute("""
             SELECT * FROM evaluacion_virtual_semana
@@ -384,7 +384,7 @@ async def calcular_resultados_virtual(
     procesados = 0
     for asig in asignaciones:
         asig_id = asig['asignacion_id']
-        horas_quincena = float(asig['horas_semana'] or 0) * n_semanas
+        horas_quincena = int(asig['horas_semana'] or 0) * n_semanas
         tarifa = float(asig['tarifa'] or 0)
         monto_base = horas_quincena * tarifa
